Move existing wasi-http tests to a `p2` module
This commit moves all existing wasi-http tests to a new `p2` module. No
actual tests are modified here, they're just moved around. This is to
make room for a future `p3` module with WASIp3-related tests.
